+++ tiny_dnn/core/kernels/fully_connected_op_internal.h
@@ -18,7 +18,7 @@ inline void fully_connected_op_internal(const tensor_t &in_data,
                                         tensor_t &out_data,
                                         const fully_params &params,
                                         const bool layer_parallelize) {
-  for_i(layer_parallelize, in_data.size(), [&](int sample) {
+  for_i(layer_parallelize, in_data.size(), [&](size_t sample) {
     const vec_t &in = in_data[sample];
     vec_t &out      = out_data[sample];
 
@@ -51,23 +51,22 @@ inline void fully_connected_op_internal(const tensor_t &prev_out,
         &curr_delta[sample][0], &W[c * params.out_size_], params.out_size_);
     }
 
-    for_(layer_parallelize, 0, size_t(params.out_size_),
-         [&](const blocked_range &r) {
-           // accumulate weight-step using delta
-           // dW[c * out_size + i] += current_delta[i] * prev_out[c]
-           for (size_t c = 0; c < params.in_size_; c++) {
-             vectorize::muladd(&curr_delta[sample][r.begin()],
-                               prev_out[sample][c], r.end() - r.begin(),
-                               &dW[sample][c * params.out_size_ + r.begin()]);
-           }
+    for_(layer_parallelize, 0, params.out_size_, [&](const blocked_range &r) {
+      // accumulate weight-step using delta
+      // dW[c * out_size + i] += current_delta[i] * prev_out[c]
+      for (size_t c = 0; c < params.in_size_; c++) {
+        vectorize::muladd(&curr_delta[sample][r.begin()], prev_out[sample][c],
+                          r.end() - r.begin(),
+                          &dW[sample][c * params.out_size_ + r.begin()]);
+      }
 
-           if (params.has_bias_) {
-             // vec_t& db = *in_grad[2];
-             for (size_t i = r.begin(); i < r.end(); i++) {
-               db[sample][i] += curr_delta[sample][i];
-             }
-           }
-         });
+      if (params.has_bias_) {
+        // vec_t& db = *in_grad[2];
+        for (size_t i = r.begin(); i < r.end(); i++) {
+          db[sample][i] += curr_delta[sample][i];
+        }
+      }
+    });
   }
 }
